fsid.jl Documentation

fsid.FRDC — Type

FRDC

Continuous-time frequency response data object with fields:

ffdata: matrix of frequency response data. ffdata[i,:,:] is the frequency function matrix (sample of the rationl function) corresponding to sample i

w: vector with angular frequencies rad/s

fsid.FRDC — Method

FRDC()

Create empty FRDC object.

fsid.FRDC — Method

FRDC(ffdata::AbstractArray, w::AbstractVector)

Create FRDC object.

fsid.FRDD — Type

FRDD

Discrete-time frequency response data object with fields:

ffdata: matrix of frequency response data. ffdata[i,:,:] is the frequency function matrix (sample of the rationl function) corresponding to sample i

z: vector with the samples of the function argument where z[i] is argument for index i

fsid.FRDD — Method

FRDD()

Create empty FRDD object.

fsid.FRDD — Method

FRDD(ffdata::AbstractArray, z::AbstractVector)

Create FRDD object.

fsid.SS — Type

SS

State-space object with fields:

a: system matrix $n \times n$

b: input matrix $n \times m$

c: output matrix $r \times n$

d: direct feedthrough matrix $r \times m$

fsid.SS — Method

SS()

Create empty SS object.

fsid.SS — Method

SS(a::AbstractMatrix, b::AbstractMatrix, c::AbstractMatrix, d::AbstractMatrix)

Create SS object.

fsid.SS — Method

SS(a::AbstractMatrix, b::AbstractMatrix, c::AbstractMatrix)

Create SS object without direct throughput.

fsid.SS — Method

SS(a::AbstractMatrix, b::AbstractMatrix)

Create SS object without output and direct throughput.

fsid.bilinear_c2d — Function

bilinear_c2d(sys::Tuple, T::Real = 1)

Calculates the bilinear transformation C->D for ss-system sys.

Parameters:

sys: tuple of system matrices (a, b, c, d)

T: frequency scaling factor

Returns:

a: system matrix

b: intput matrix

c: output matrix

d: direct feedthrough matrix

fsid.bilinear_d2c — Function

bilinear_d2c(sys::Tuple, T::Real = 1)

Calculates the bilinear transformation D->C for state-space system sys.

Parameters:

sys: tuple of system matrices (a, b, c, d)

T: frequency scaling factor

Returns:

a: system matrix

b: input matrix

c: output matrix

d: direct feedthrough matrix

fsid.cf2df — Method

cf2df(wc::AbstractVector{<:Number}, T::Real)

Calculates the bilinear transformation frequency mapping C->D for frequency vector wc.

Parameters:

wc: vector of frequencies

T: frequency scaling factor

Returns:

wd: vector of transformed frequencies

fsid.contype — Function

contype(type::Type{<:Real}, dtype = Float64)
contype(type::Type{<:Complex}, dtype = Float64)

Convert type to default concrete type.

Parameters:

type: a concrete or abstract type

Returns:

default concrete type Float64

fsid.df2cf — Method

df2cf(wd::AbstractVector{<:Number}, T::Real)

Calculates the bilinear transformation frequency mapping D->C for frequency vector wd.

Parameters:

wd: vector of frequencies

T: frequency scaling factor

Returns:

wc: vector of transformed frequencies

fsid.estimate_bd — Method

estimate_bd(
    ffdata::AbstractArray{<:Number, 3},
    z::AbstractVector{<:Number},
    a::AbstractMatrix{<:Number},
    c::AbstractMatrix{<:Number};
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true,
)

Estimates the b and d matrices given a, c matrices and frequency function data ffdata.

Solve the optimization problem

$\min_{b,d} \sum_i \| d + c*inv(z[i]*I-a)*b - ffdata[i,:,:] \|^2_F$

or if estimd == false

$\min_{b} \sum_i \| c*inv(z[i]*I-a)*b - ffdata[i,:,:] \|^2_F$

if type = Real a real valued solution is calulated. If type = Complex the solution is complex valued

Parameters:

ffdata: frequency data packed in a matrix. ffdata[i,:,:] is the frequency function matrix corresponding to sample i

z: vector with complex scalars

a: square matrix

c: matrix

Optional_

type: data type of model, either Real or Complex

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

Returns:

b: LS-optimal b matrix

d: LS-optimal d matrix

fsid.estimate_cd — Method

estimate_cd(
    ffdata::AbstractArray{<:Number, 3},
    z::AbstractVector{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number};
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true,
)

Estimates the c and d matrices given a andbmatrices and frequency function dataffdata`.

Solve the optimization problem

$\min_{c,d} \sum_i \| d + c*inv(z[i]*I-a)*b - ffdata[i,:,:] \|^2_F$

or if estimd == false

$\min_{c} \sum_i \| c*inv(z[i]*I-a)*b - ffdata[i,:,:] \|^2_F$

if type = Real a real valued solution is calulated. If type = Complex the solution is complex valued

Parameters:

ffdata: frequency data packed in a matrix. ffdata[i,:,:] is the frequency function matrix corresponding to sample i

z: vector with complex scalars

a: square matrix

b: matrix

Optional:

type: data type of model, either Real or Complex

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

Returns :

c: LS-optimal c matrix

d: LS-optimal d matrix

fsid.fdestim_bd — Method

fdestim_bd(
    z::AbstractVector{<:Number},
    yd::AbstractMatrix{<:Number},
    ud::AbstractMatrix{<:Number},
    a::AbstractMatrix{<:Number},
    c::AbstractMatrix{<:Number};
    estTrans::Bool = false,
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true
)

Estimate b and d matrices (and optionally xt) given yd, ud, a and c matrices.

Calulates the b and d matrices (and optimally xt) for a linear dynamic system in state-space form given the a and c matrices and samples of frequency domain function data. It solves if estimd = true and estTrans = true

$\min_{b,d,xt} \sum_i \| ([d 0] + c*inv(z[i]*I-a)*[b xt])[ud[i,:]; z[i]] - ffdata[i,:,:] \|^2_F$

if estimd = false and estTrans = true

$\min_{b,xt} \sum_i \| (c*inv(z[i]*I-a)*[b xt])[ud[i,:]; w[i]] - ffdata[i,:,:] \|^2_F$

if estimd = true and estTrans = false

$\min_{b,d} \sum_i \| (d+ c*inv(z[i]*I-a)*b)ud[i,:] - ffdata[i,:,:] \|^2_F$

if estimd = false and estTrans = false

$\min_{b} \sum_i \| (c*inv(z[i]*I-a)*b)ud[i,:] - ffdata[i,:,:] \|^2_F$

if type = Real a real valued solution is calulated. if type = Complex the solution is complex valued

Parameters:

z: vector with the samples of the function argument where z[i] is argument for index i

yd: output frequency data yd[i,:]

ud: input frequency data ud[widx,:]

a: system matrix

c: output matrix

Optional:

estTrans: if set to true also an xt vector will be estimated capturing the transient effect

type: data type of model, either Real or Complex

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

Returns:

b: the LS-optimal b matrix

d: LS-optimal d matrix zeros matrix if estimd = false

xt: LS-optimal xt vector if estTrans = true

fsid.fdestim_cd — Function

fdestim_cd(
    z::AbstractVector{<:Number},
    yd::AbstractArray{<:Number},
    ud::AbstractArray{<:Number},
    a::AbstractArray{<:Number},
    b::AbstractArray{<:Number},
    xt::Union{AbstractVector{<:Number}, Number} = 0;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true
)

Estimate c and d matrices given z, yd, ud and a, c, and (optionally xt) matrices.

Calulates the c and d matrices for a linear dynamic system on state-space form given the a and b matrices and samples of rational function data. It solves if estimd = true

$\min_{c,d} \sum_i \| ([d 0] + c*inv(z[i]*I-a)*[b xt])[ud[i,:]; z[i]] - ffdata[i,:,:] \|^2_F$

if estimd = false

$\min_{c} \sum_i \| (c*inv(z[i]*I-a)*[b xt])[ud[i,:]; w[i]] - ffdata[i,:,:] \|^2_F$

if type = Real a real valued solution is calulated. if type = Complex the solution is complex valued

Parameters:

ffdata: data packed in a matrix. ffdata[i,:,:] is the frequency function matrix (sample of the rationl function) corresponding to sample i

w: vector with the rational function arguments. w[i] is the function arguemnt for sample i

a: matrix

b: matrix

Optional:

xt: vector

type: data type of model either Real or Complex

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

Returns :

c: LS-optimal c matrix

d: LS-optimal d matrix

fsid.fdsid — Method

function fdsid(
    fddata::Tuple,
    n::Integer,
    q::Integer;
    estTrans::Bool = true,
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true,
    CT::Bool = false,
    T::Real = 1
)

Estimate a DT or CT state-space model from I/O frequency data.

Determines the (a,b,c,d,xt) parametrers such that (DT case)

$sum_i || y[i,:] - d*u[i, :] + c*inv(z[i]*eye(n)-a)*[b, xt]* [u[i, :]; z[i]]||^_F$

is small where z[i] = exp(im*w[i]) and CT Case

$sum_i || y[i,:] - d*u[i, :] + c*inv(im*w[i]*eye(n)-a)*b* u[i, :] ||^2_F$

Parameters:

fddata: a tuple with elements

fddata[0] = w: a vector of frequencies in radians/sample (rad/s in CT case)

fddata[1] = y: a matrix of the output transform data where y[i,:] is the DFT of the time domain output signal corresponding to frequency w[i]

fddata[2] = u: a matrix of the input transform data where u[i,:] is the DFT of the time domain input signal corresponding to frequency w[i]

n: the model order of the ss-model

q: the numer of block rows used in the intermediate matrix. Must satisfy q > n

Optional:

estTrans: if true, a compensation for the transient term will be estimated (default)

type:

if type = Real a real valued solution (a, b, c, d) is returned.

if type = Complex a complex valued solution is returned.

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

CT: if true a CT model is estimated and estTrans is forced false

T: a frequency scaling factor for the bilinear transformation used when CT = true. Default is 1. If CT = false parameter T is disregarded

Returns:

a: the estimated a matrix

b: the estimated b matrix

c: the estimated c matrix

d: the estimated d matrix (or zero matrix if estimd = false)

xt: vector of the transient compensation

s: a vector of the singular values

fsid.fdsim — Function

function fdsim(
    sys::Tuple,
    u::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    xt::AbstractMatrix{<:Number} = Matrix{Float64}(undef, 0, 0)
)

Calculates the output given input and state-space model in Fourier domain

$x[i,:] = inv(z[i]*I-a)*[B xt]*[u[i,:]; z[i]]$

$y[i,:] = d*u[i,:] + c*x[i,:]$

Parameters:

sys: typle sys = (a, b, c, d) or sys = (a, b, c) where a is a square matrix of size (n,n), b is a matrix of size (n, m), c is a matrix of size (p, n) and (optionally) d is a matrix of size (p, m).

u: an array of input vectors such that u[i,:] is the input vector of size m at sample index i. The array has size (N, m)

z: vector with the samples of the frequency function argument

xt: transient vector of size n, Default value is the zero vector.

Returns:

y: the resulting output sequence of size (N,p)

x: the resulting state sequence of size (N,p)

fsid.ffdata2fddata — Method

ffdata2fddata(ffdata::AbstractArray{<:Number, 3}, z::AbstractVector{<:Number})

Converts ffdata to fddata.

Converts frequency function data ffdata to input/output data format fddata.

Parameters:

ffdata: frequency function data in the format such that ffdata[i,:,:] corresponds to the frequency function matrix of size (p,m) at frequency index i corresponding to function argument z[i] at a total number of samples nz

z: array with the corresponding frequency function argument of size nz

Returns:

u: Fourier transform of input of size (m*nz, m)

y: Fourier transform of output of size (m*nz, p)

zn: frequency function argument vector of length m*nw

fsid.ffsid — Method

function ffsid(
    w::AbstractVector{<:Number},
    ffdata::AbstractArray{<:Number, 3},
    n::Integer,
    q::Integer;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true,
    CT::Bool = false,
    T::Real = 1
)

Estimate a state-space model (a, b, c, d) from frequency function data.

Parameters:

w: vector of frequencies in rad/sample [-pi,pi] or [0, 2pi]. if CT = true unit in radians/s (-inf, +inf)

ffdata: matrix of frequency function data. ffdata[i,:,:] is the frequency response matrix at frequency w[i]

n: the model order of the ss-model (a, b, c, d), i.e. a is a size (n x n) matrix

q: the numer of block rows used in the intermediate matrix. Must satisfy q > n

Optional:

type:

if type = Real a real valued solution (a, b, c, d) is returned

if type = Complex a complex valued solution (a,b,c,d) is returned

estimd: if set to False no d matrix is esimated and a zero d matrix is returned

CT:

if set to true a continuous time (CT) model is esimated

if set to false a discrete time (DT) model is esimated (default)

T: a frequency scaling factor for the bilinear transformation used when CT = true. Default is 1. If CT = false parameter T is disregarded

Returns:

a: the estimated a matrix

b: the estimated b matrix

c: the estimated c matrix

d: the estimated d matrix (or zero matrix if estimd = false)

s: a vector of the singular values

fsid.fresp — Function

function fresp(
    z::AbstractVector{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    c::AbstractMatrix{<:Number},
    d::AbstractMatrix{<:Number},
    noWarning::Bool = true
)

Frequency response of ss-model (a,b,c,d) a rational matrix function

$fresp[i,:,:] = d+c*inv(z[i]*I-a)*b$

fresp[i,:,:] is the function value of the rational matrix function evaluated at z[i]

Parameters:

z: vector with samples of the function argument

a: matrix

b: matrix

c: matrix

d: matrix

noWarning: information message is suppressed if set to true

Returns:

frsp: frequency response

fsid.fresp_fast! — Method

function fresp_fast!(
   frsp::AbstractArray{<:Number, 3},
    z::AbstractVector{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    c::AbstractMatrix{<:Number},
    d::AbstractMatrix{<:Number},
    eig::Eigen
)

Frequency response of ss-model (a, b, c, d) (fast version)

$fresp[i,:,:] = d+c*inv(I*z[i]-a)*b$

fsid.fresp_slow! — Method

function fresp_slow!(
    frsp::AbstractArray{<:Number, 3},
    z::AbstractVector{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    c::AbstractMatrix{<:Number},
    d::AbstractMatrix{<:Number}
)

Frequency response of state-space model (a, b, c, d) (slow version) calulated as

$fresp[i,:,:] = d + c * (I*z[i]-a)^{-1} * b$

fsid.gfdsid — Method

function gfdsid(
    fddata::Tuple,
    n::Integer,
    q::Integer;
    estTrans::Bool = true,
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true
)

Estimate a state-space model from I/O frequency data

Determines the (a, b, c, d, xt) parameters such that

$\sum_i \| y[i,:] - d*u[i, :] + c*inv(z[i]*eye(n)-a)*[b, xt]* [u[i, :]; z[i]] \|^2_F$

is minimized.

If estrTrans = false the following problem is solved

$\sum_i \| y[i,:] - d*u[i, :] + c*inv(z[i]*eye(n)-a)*b * u[i, :] \|^2_F$

is minimized.

Parameters:

fddata: a tuple with elements

fddata[0] = z: a vector of complex scalars,

fddata[1] = y: a matrix of the output frequency data where y[i,:] corresponds to z[i]

fddata[2] = u: a matrix of the input frrequency data where u[i,:] corresponding to z[i]

n: the model order of the ss-model

q: the numer of block rows used in the intermediate matrix. Must satisfy q > n

Optional:

estTrans: if true, a compensation for the transient term will be estimated (default)

type:

if type = Real a real valued solution (a, b, c, d) is returned.

if type = Complex a complex valued solution is returned.

estimd: if set to 'false' no d matrix is esimated and a zero d matrix is returned

Returns:

a: the estimated a matrix

b: the estimated b matrix

c: the estimated c matrix

d: the estimated d matrix (or zero matrix if estimd = false)

x: vector of the transient compensation

s: a vector of the singular values

fsid.gffsid — Method

function gffsid(
    z::AbstractVector{<:Number},
    ffdata::AbstractArray{<:Number, 3},
    n::Integer,
    q::Integer;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real,
    estimd::Bool = true
)

Estimate a state-space model (a, b, c, d) from general frequency function data by mininizing the Frobenius norm

$\sum_i \| ffdata[i,:,:] - d - c*inv(z[i]*eye(n)-a)*b \|^2_F$

Parameters:

z: vector of complex numbers

ffdata: matrix of frequency function data. ffdata[i,:,:] is the frequency response matrix at complex point z[i]

n: the model order of the ss-model (a, b, c, d), i.e. a is a size (n x n) matrix

q: the numer of block rows used in the intermediate matrix. Must satisfy q > n

Optional:

type:

if type = Real a real valued solution (a, b, c, d) is returned

if type = Complex a complex valued solution is returned

estimd: if set to false no d matrix is esimated and a zero d matrix is returned

Returns:

a: the estimated a matrix

b: the estimated b matrix

c: the estimated c matrix

d: the estimated d matrix (or zero matrix if estimd = false)

s: a vector of the singular values

fsid.lsim — Function

function lsim(
    sys::Tuple,
    u::AbstractMatrix{<:Number},
    x0::Union{AbstractVector{<:Number}, Number} = 0;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real
)

Calculates the time-domain output given input sequence and state-space model sys = (a, b, c, d)

$x[i+1,:] = a*x[i,:]) + b*u[i,:]$

$y[i,:] = c*x[i,:]) + d*u[i,:]$

Parameters:

sys: a typle sys = (a, b, c, d) or sys = (a, b, c) where a is a square matrix of size (n,n), b is a matrix of size (n,m), c is a matrix of size (p,n) and (optionally) d is a matrix of size (p, m)`

u: an array of input vectors such that u[i,:] is the input vector of size m at time index i. The array has size (N, m)

Optional:

x0: intial vector of size n, i.e. x[0,:] = x0. Default value is the zero vector

Returns:

y: the resulting output sequence of size (N, p)

x: the resulting state sequence of size (N, n)

fsid.lsim — Function

function lsim(
    sys::Tuple,
    u::AbstractMatrix{<:Number},
    x0::Union{AbstractVector{<:Number}, Number} = 0;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real
)

Calculates the time-domain output given input sequence and state-space model sys = (a, b, c, d)

$x[i+1,:] = a*x[i,:]) + b*u[i,:]$

$y[i,:] = c*x[i,:]) + d*u[i,:]$

Parameters:

sys: a typle sys = (a, b, c, d) or sys = (a, b, c) where a is a square matrix of size (n,n), b is a matrix of size (n,m), c is a matrix of size (p,n) and (optionally) d is a matrix of size (p,m)

u: an array of input vectors such that u[i,:] is the input vector of size m at time index i. The array has size (N,m).

Optional:

x0: intial vector of size n, i.e. x[0,:] = x0. Default value is the zero vector

Returns:

y: the resulting output sequence of size (N, p)

x: the resulting state sequence of size (N, n)

fsid.ltifd — Function

function ltifd(
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    u::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    noWarning::Bool = true
)

Calculates the frequency kernel with input u

$fkern[:, i] = inv(z[i]*I - a)*b*u[i, :]$

Parameters:

a: matrix

b: matrix

u: input vectors

z: vector with samples of the function argument

noWarning: if true suppress information message

Returns:

fkern: frequency kernel

fsid.ltifd_fast! — Method

function ltifd_fast!(
    fkern::AbstractMatrix{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    u::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    eig::Eigen
)

Calculates the frequency kernel with input u in place (fast version)

$fkern[:, i] = inv(z[i]*I - a)*b*u[i, :]$

Parameters:

fkern: frequency kernel

a: matrix

b: matrix

u: matrix with input vectors

z: vector with the frequecy data

eig: eigendecomposition of a

fsid.ltifd_slow! — Method

function ltifd_slow!(
    fkern::AbstractMatrix{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    u::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number}
)

Calculates the frequency kernel with input u in place (slow version)

$fkern[:, i] = inv(z[i]*I - a)*b*u[i, :]$

Parameters:

fkern: frequency kernel

a: matrix

b: matrix

u: matrix with input vectors

z: vector with the frequecy data

fsid.ltifd_slow! — Method

function ltifd_slow!(
    fkern::AbstractMatrix{<:Number},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    u::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number}
)

Calculates the frequency kernel with input u in place (slow version)

$fkern[:, i] = inv(z[i]*I - a)*b*u[i, :]$

Parameters:

fkern: frequency kernel

a: matrix

b: matrix

u: matrix with input vectors

z: vector with the frequecy data

fsid.ltifr — Function

function ltifr(
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    noWarning::Bool = true    
)

Calculates the frequency kernel

$fkern[i,:,:] = inv(z[i]*I-a)*b$

Parameters:

a: matrix

b: matrix

z: vector with the samples of the function argument

noWarning: if true suppress information message

Returns:

fkern: frequency response

fsid.ltifr_fast! — Method

function ltifr_fast!(
    fkern::AbstractArray{<:Number, 3},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    eig::Eigen
)

Calculates the frequency kernel in place (fast version)

$fkern[i,:,:] = inv(z[i]*I-a)*b$

fsid.ltifr_fast! — Method

function ltifr_fast!(
    fkern::AbstractArray{<:Number, 3},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number},
    eig::Eigen
)

Calculates the frequency kernel in place (fast version)

$fkern[i,:,:] = inv(z[i]*I-a)*b (slow version)$

fsid.ltifr_slow! — Method

function ltifr_slow!(
    fkern::AbstractArray{<:Number, 3},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number}
)

Calculates the frequency kernel in place as (slow version)

$fkern[i,:,:] = inv(z[i]*I-a)*b$

fsid.ltifr_slow! — Method

function ltifr_slow!(
    fkern::AbstractArray{<:Number, 3},
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    z::AbstractVector{<:Number}
)

Calculates the frequency kernel in place as

$fkern[i,:,:] = inv(z[i]*I-a)*b (slow version)$

fsid.ltitr — Function

function ltitr(
    a::AbstractMatrix{<:Number},
    b::AbstractMatrix{<:Number},
    u::AbstractMatrix{<:Number},
    x0::Union{AbstractVector{<:Number}, Number} = 0;
    type::Union{Type{<:Real}, Type{<:Complex}} = Real
)

Calculates the time domain input to state respone.

Calculates the time domain state response

$x[i+1,:] = a*x[i,:]) + b*u[i,:]$

Parmeters:

a: a square matrix of size (n, n)

b: a matrix of size (n, m)

u: an array of input vectors such that u[i,:] is the input vector of size m at time index i. The array has size (N, m)

x0: intial vector of size n, i.e. x[0,:] = x0. Default value is the zero vector

Returns:

x: the resulting state-sequence of size (N, n)

x[k,:]: is the state at sample k

fsid.matrix_chain_order — Method

matrix_chain_order(p)

Reference: Cormen, "Introduction to Algorithms", Chapter 15.2, p. 370-378

fsid.multidot — Method

multidot(mats...)

Fast multiplication of multiple matrices, i.e. optimal order. Provide matrices in order of multiplication.

fsid.optimal_print — Method

optimal_print(s, i, j)

Helper function.

fsid.transpose_ffdata — Method

transpose_ffdata(ffdata::AbstractArray{<:Number, 3})

Transposes ffdata (changes inputs and outputs).